Revert "sizes passed to the CefRenderHandler OnPaint and OnPopupSize methods are now already DPI scaled"
This reverts commit d173f28e0900840ef6466ccf22808467faa23b10.

+++ CefSharp/Internals/IRenderWebBrowser.cs
@@ -10,8 +10,8 @@ namespace CefSharp.Internals
     {
         int BytesPerPixel { get; }
 
-        double Width { get; }
-        double Height { get; }
+        int Width { get; }
+        int Height { get; }
 
         void InvokeRenderAsync(BitmapInfo bitmapInfo);
